Moving the build to CBI-signing
maven.eclipse.org is now down and the plugin we used previously
for the signing is no more available.
This commit (tentatively) upgrade the build to use cbi signing plugin
The major change I noted is that instead of configuring it only on the
update-site pom.xml, it has to be configured in a parent pom and then
every jar which is produced is going to be signed.

This plugin also no longer needs a specific server path for signing.

Signing is only done when the build-server profile is enabled.